chore: move new version prompt version to action button
