+++ web/src/ee/features/playground/page/components/JumpToPlaygroundButton.tsx
@@ -62,13 +62,16 @@ export const JumpToPlaygroundButton: React.FC<JumpToPlaygroundButtonProps> = (
   return (
     <Button
       variant={props.variant ?? "secondary"}
+      size={props.source === "prompt" ? "icon" : "default"}
       title="Test in LLM playground"
       onClick={handleClick}
       asChild
     >
       <Link href={`/project/${projectId}/playground`}>
         <Terminal className="h-4 w-4" />
-        <span className="ml-2">Test in playground</span>
+        {props.source === "generation" && (
+          <span className="ml-2">Test in playground</span>
+        )}
       </Link>
     </Button>
   );
