+++ web/src/features/prompts/components/prompt-detail.tsx
@@ -25,14 +25,23 @@ import {
 import { JumpToPlaygroundButton } from "@/src/ee/features/playground/page/components/JumpToPlaygroundButton";
 import { ChatMlArraySchema } from "@/src/components/schemas/ChatMlSchema";
 import { CommentList } from "@/src/features/comments/CommentList";
+import { Lock, Plus } from "lucide-react";
+import { useHasProjectAccess } from "@/src/features/rbac/utils/checkProjectAccess";
+import { Button } from "@/src/components/ui/button";
+import { usePostHogClientCapture } from "@/src/features/posthog-analytics/usePostHogClientCapture";
 
 export const PromptDetail = () => {
   const projectId = useProjectIdFromURL();
+  const capture = usePostHogClientCapture();
   const promptName = decodeURIComponent(useRouter().query.promptName as string);
   const [currentPromptVersion, setCurrentPromptVersion] = useQueryParam(
     "version",
     NumberParam,
   );
+  const hasAccess = useHasProjectAccess({
+    projectId,
+    scope: "prompts:CUD",
+  });
   const promptHistory = api.prompts.allVersions.useQuery(
     {
       name: promptName,
@@ -49,7 +58,7 @@ export const PromptDetail = () => {
   const extractedVariables = prompt
     ? extractVariables(
         prompt?.type === PromptType.Text
-          ? prompt.prompt?.toString() ?? ""
+          ? (prompt.prompt?.toString() ?? "")
           : JSON.stringify(prompt.prompt),
       )
     : [];
@@ -105,6 +114,27 @@ export const PromptDetail = () => {
             ]}
             actionButtons={
               <>
+                {hasAccess ? (
+                  <Button
+                    variant="outline"
+                    onClick={() => {
+                      capture("prompts:update_form_open");
+                    }}
+                  >
+                    <Link
+                      href={`/project/${projectId}/prompts/new?promptId=${encodeURIComponent(prompt.id)}`}
+                    >
+                      <div className="flex flex-row items-center">
+                        <Plus className="h-4 w-4" />
+                        <span className="ml-2">New version</span>
+                      </div>
+                    </Link>
+                  </Button>
+                ) : (
+                  <Button variant="outline" disabled>
+                    <Lock className="h-4 w-4" />
+                  </Button>
+                )}
                 <JumpToPlaygroundButton
                   source="prompt"
                   prompt={prompt}
